+++ bridges/otellogrus/hook.go
@@ -33,9 +33,6 @@
 package otellogrus // import "go.opentelemetry.io/contrib/bridges/otellogrus"
 
 import (
-	"fmt"
-	"reflect"
-
 	"github.com/sirupsen/logrus"
 
 	"go.opentelemetry.io/otel/log"
@@ -177,7 +174,7 @@ func convertFields(fields logrus.Fields) []log.KeyValue {
 	for k, v := range fields {
 		kvs = append(kvs, log.KeyValue{
 			Key:   k,
-			Value: convertValue(v),
+			Value: ConvertValue(v),
 		})
 	}
 	return kvs
@@ -205,57 +202,4 @@ func convertSeverity(level logrus.Level) log.Severity {
 		// we should never reach this point as logrus only uses the above levels.
 		return log.SeverityUndefined
 	}
-}
-
-func convertValue(v interface{}) log.Value {
-	switch v := v.(type) {
-	case bool:
-		return log.BoolValue(v)
-	case []byte:
-		return log.BytesValue(v)
-	case float64:
-		return log.Float64Value(v)
-	case int:
-		return log.IntValue(v)
-	case int64:
-		return log.Int64Value(v)
-	case string:
-		return log.StringValue(v)
-	}
-
-	t := reflect.TypeOf(v)
-	if t == nil {
-		return log.Value{}
-	}
-	val := reflect.ValueOf(v)
-	switch t.Kind() {
-	case reflect.Struct:
-		return log.StringValue(fmt.Sprintf("%+v", v))
-	case reflect.Slice, reflect.Array:
-		items := make([]log.Value, 0, val.Len())
-		for i := 0; i < val.Len(); i++ {
-			items = append(items, convertValue(val.Index(i).Interface()))
-		}
-		return log.SliceValue(items...)
-	case reflect.Map:
-		kvs := make([]log.KeyValue, 0, val.Len())
-		for _, k := range val.MapKeys() {
-			var key string
-			// If the key is a struct, use %+v to print the struct fields.
-			if k.Kind() == reflect.Struct {
-				key = fmt.Sprintf("%+v", k.Interface())
-			} else {
-				key = fmt.Sprintf("%v", k.Interface())
-			}
-			kvs = append(kvs, log.KeyValue{
-				Key:   key,
-				Value: convertValue(val.MapIndex(k).Interface()),
-			})
-		}
-		return log.MapValue(kvs...)
-	case reflect.Ptr, reflect.Interface:
-		return convertValue(val.Elem().Interface())
-	}
-
-	return log.StringValue(fmt.Sprintf("unhandled attribute type: (%s) %+v", t, v))
 }